asfdhf

+++ Network/Dispatcher.h
@@ -8,8 +8,8 @@ private:
 	static int CurrentTime;
 public:
 	static void QueuePacketCreation(int nodeId, int destinationId);
-	static void QueuePacketProcessing(int nodeId, std::shared_ptr<Packet> packet, double delay);
-	static void QueuePacketUpload (int linkId, int sourceId, std::shared_ptr<Packet> packet, double delay);
-	static void QueuePacketDownload(int linkId, std::shared_ptr<Packet> packet, double delay);
+	static void QueuePacketProcessing(int nodeId, std::shared_ptr<Packet> packet, int delay);
+	static void QueuePacketUpload (int linkId, int sourceId, std::shared_ptr<Packet> packet, int delay);
+	static void QueuePacketDownload(int linkId, std::shared_ptr<Packet> packet, int delay);
 	static int GetCurrentTime();
 };
